requirements.txt
Flask==1.0.2
line-bot-sdk==1.7.2